% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPlus.R
\name{getPsi}
\alias{getPsi}
\title{Compute the Psi design matrix}
\usage{
getPsi(x, ...)
}
\arguments{
\item{x}{Object to dispatch on.}

\item{...}{Further arguments passed to methods.}
}
\description{
Compute the Psi design matrix
}
