% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFplus.R
\name{RFplus}
\alias{RFplus}
\title{RFplus: Bias correction model using Random Forest for satellite-derived environmental data.}
\usage{
RFplus(
  Covariates,
  BD.Insitu,
  Cords_Insitu,
  ntree = 2000,
  threshold = NULL,
  save_model = F,
  name_save = NULL
)
}
\arguments{
\item{Covariates}{A list of covariates used as independent variables.
The selected covariates should be provided as a list of \code{SpatRaster} objects.
Each covariate must be stored in a separate object within the list.
Additionally, make sure to include the DEM (Digital Elevation Model).}

\item{BD.Insitu}{A \code{data.table} file containing the in-situ station coordinates.
This file must have 3 columns: the first column is \code{Cod}, which indicates the unique code
for identifying each station. This code should match the one used in the \code{data.table} of the 'BD.Insitu'.
The second and third columns represent the \code{X} and \code{Y} coordinates of each in-situ station, respectively.}

\item{ntree}{An integer representing the number of trees to use in the random forest model. Default is 2000.}

\item{threshold}{A numeric value for setting a threshold for the final predictions. Values below this threshold will be set to 0. Default is NULL.}

\item{save_model}{A logical value indicating whether to save the final corrected model in NetCDF format.
Default is \code{FALSE}, which does not save the model.
If \code{TRUE}, the corrected model will be saved to a NetCDF file."
It is important that, before executing the function, the directory where the file will be saved is defined.
using \code{setwd(“path/directory”)}.}

\item{name_save}{The name under which the model will be saved if \code{save_model = TRUE}.
By default, the model is saved with the name \code{"Model_RFplus"}.
The name should be provided without an extension, as the function will append the appropriate file extension (.nc).}
}
\value{
A \code{terra} raster object containing the corrected environmental data.
        The output will have the same spatial resolution and extent as the input covariates.
}
\description{
This function applies a Random Forest-based bias correction method to satellite-derived environmental data, such as precipitation, temperature, and other relevant meteorological or environmental variables.
The model corrects biases in these satellite products by incorporating in-situ measurements and environmental covariates, including the Digital Elevation Model (DEM).
}
\author{
Jonnathan Augusto Landi Bermeo
}
